--- constants/colors.js
@@ -1,8 +1,0 @@
-// export const colors = [
-//     {
-//         name:
-//     }
-
-// ];
-
-// This file will contain arrays of all pastel and dark colors
