+++ constants/colors.jsx
@@ -1,0 +1,7 @@
+export const cardColors = [
+  { pastel: '#FCF0E3', dark: '#BA7D37' },
+  { pastel: '#EBF9FA', dark: '#5B8A8D' },
+  { pastel: '#EFFAEB', dark: '#75946A' },
+];
+
+// This file will contain arrays of all pastel and dark colors
