+++ constants/mockData.jsx
@@ -1,0 +1,71 @@
+export const professorInfoMockData = {
+  featuredCourses: [{
+    courseNumber: 'COSC 52',
+    courseName: 'Full Stack Web Development',
+    term: '21F',
+    quality: '4.0',
+    difficulty: '3.0 (3)',
+    hrsPerWeek: '3.0 (3)',
+    median: 'A-',
+    reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+    distribs: ['TLA', 'NW'],
+  },
+  {
+    courseNumber: 'COSC 98.01',
+    courseName: 'Senior Design and Implementation I',
+    term: '19F',
+    quality: '5.0',
+    difficulty: '4.0 (3)',
+    hrsPerWeek: '3.0 (4)',
+    median: 'A',
+    reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+    distribs: ['TLA', 'NW'],
+  },
+  {
+    courseNumber: 'COSC 98.02',
+    courseName: 'Senior Design and Implementation II',
+    term: '19W',
+    quality: '5.0',
+    difficulty: '5.0 (3)',
+    hrsPerWeek: '3.0 (3)',
+    median: 'A',
+    reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+    distribs: ['TLA', 'NW'],
+  },
+  ],
+  allCourses: [
+    {
+      courseNumber: 'COSC 52',
+      courseName: 'Full Stack Web Development',
+      term: '21F',
+      quality: '4.0',
+      difficulty: '3.0 (3)',
+      hrsPerWeek: '3.0 (3)',
+      median: 'A-',
+      reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+      distribs: ['TLA', 'NW'],
+    },
+    {
+      courseNumber: 'COSC 98.01',
+      courseName: 'Senior Design and Implementation I',
+      term: '19F',
+      quality: '5.0',
+      difficulty: '4.0 (3)',
+      hrsPerWeek: '3.0 (4)',
+      median: 'A',
+      reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+      distribs: ['TLA', 'NW'],
+    },
+    {
+      courseNumber: 'COSC 98.02',
+      courseName: 'Senior Design and Implementation II',
+      term: '19W',
+      quality: '5.0',
+      difficulty: '5.0 (3)',
+      hrsPerWeek: '3.0 (3)',
+      median: 'A',
+      reviews: ['This class was awesome', 'Tim is great', 'Natalie is amazing'],
+      distribs: ['TLA', 'NW'],
+    },
+  ],
+};
