+++ styles/Social.module.css
@@ -2,6 +2,7 @@
   padding: 0 2rem;
   margin-left: 250px;
   margin-top: 2%;
+  flex-direction: row;
   /* height: 100%; */
 }
 
@@ -18,4 +19,12 @@
   width: 100%;
   margin-top: 3%;
   margin-bottom: 3%;
+}
+
+.coursesContainer {
+  display: flex;
+  flex-direction: row;
+  margin-top: 10px;
+  margin-bottom: 10px;
+  flex-wrap: wrap;
 }
